Add files via upload
cropgraphprecise.java:  This is the code that I used to obtain the best optimizations for first place in this category.
optimization2.java: A test run I did that would also locate the smallest circle around a group of points. It did not affect the final score and increased time complexity, so it was not used in the end.